Изменен класс UserService, добавлены ключи-зависимости в таблицах
