+++ src/main/java/com/example/shelter_bot/entity/Client.java
@@ -1,10 +1,10 @@
 package com.example.shelter_bot.entity;
 
-import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import javax.persistence.*;
 import java.util.HashSet;
 import java.util.Set;
 
@@ -17,7 +17,7 @@ public class Client {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-    @Column(nullable = false)
+    @Column(name = "name_client", nullable = false)
     private String name;
     @Column(name = "phone_number", nullable = false)
     private String phoneNumber;
