+++ src/main/java/com/example/shelter_bot/entity/User.java
@@ -1,9 +1,10 @@
 package com.example.shelter_bot.entity;
 
-import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
+
+import javax.persistence.*;
 
 @Entity
 @Table(name = "users")
@@ -14,13 +15,15 @@ public class User {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-    @Column(nullable = false)
+    @Column(name = "name_user", nullable = false)
     private String name;
     @Column(name = "phone_number", nullable = false)
     private String phoneNumber;
     @ManyToOne
     @JoinColumn(name = "shelter_id")
     private Shelter shelter;
+    @Column(name = "user_chat_id")
+    private Long userChatId;
 
     public User(String name, String phoneNumber, Shelter shelter) {
         this.name = name;
