+++ src/main/resources/liquibase/init.sql
@@ -2,22 +2,22 @@ CREATE TABLE pet
 (
     id                     BIGINT generated by default as identity primary key,
     pet_type               TEXT      NOT NULL,
-    chat_id                BIGINT    NOT NULL,
+    chat_id                BIGINT    NOT NULL REFERENCES users (user_chat_id)
 );
 
 CREATE TABLE client
 (
     id                     BIGINT generated by default as identity primary key,
-    name                   TEXT      NOT NULL,
+    name_client            TEXT      NOT NULL,
     phoneNumber            INT       NOT NULL,
     address                TEXT      NOT NULL,
-    pet_set                TEXT      NOT NULL,
+    pet_set                TEXT      NOT NULL
 );
 
 CREATE TABLE shelter
 (
     id                     BIGINT generated by default as identity primary key,
-    name                   TEXT      NOT NULL,
+    name_shelter           TEXT      NOT NULL,
     address                TEXT      NOT NULL,
     schedule               INT       NOT NULL,
     about                  TEXT      NOT NULL,
@@ -25,23 +25,24 @@ CREATE TABLE shelter
     location_map           TEXT      NOT NULL,
     pet_type               TEXT      NOT NULL,
     volunteerSet           TEXT      NOT NULL,
-    ownerSet               TEXT      NOT NULL,
+    ownerSet               TEXT      NOT NULL
 );
 
 CREATE TABLE users
 (
     id                     BIGINT generated by default as identity primary key,
-    name                   TEXT      NOT NULL,
+    name_user              TEXT      NOT NULL,
     phoneNumber            INT       NOT NULL,
-    shelter                TEXT      NOT NULL,
+    shelter_id             INT       NOT NULL REFERENCES shelter (id),
+    user_chat_id           BIGINT NOT NULL UNIQUE
 );
 
 CREATE TABLE volunteers
 (
     id                     BIGINT generated by default as identity primary key,
-    name                   TEXT      NOT NULL,
-    chat_id                BIGINT    NOT NULL,
-    availability           BOOL      NOT NULL,
+    name_volunteer         TEXT      NOT NULL,
+    chat_id                BIGINT    NOT NULL REFERENCES users (user_chat_id),
+    available              BOOL      NOT NULL
 
 );
